refactor(settings): move Background Agent picker to the Agents tab
The dropdown lived in Settings → Extensions next to the OpenCode/Hostler
provider cards, but the Agents tab is where agent behavior is configured
(authentication, browser automation, MCP servers) — the picker belongs
there. It now sits at the top of Agent Settings.

Provider gates derive from the react-query general-config data instead of
copied state, and the config refetches whenever the Agents tab is shown so
enablement toggled in the Extensions tab is reflected without reopening
Settings. Verified live over CDP: option gating, selection persistence,
fallback warning after disabling Hostler from another tab, and the card's
removal from Extensions. PR screenshots updated.
